Write first user documentation chapters
With this we transition away from a training-based introduction to a more standardised
read the docs setup, that provides references for different types of users. What is meant
to follow this is a section containing developer or maintainer documentation.

Parts of the getting started chapters were modified for flow and incremental introduction.